rename SnapshotX to SegmentX - rename files
