+++ silkworm/db/blocks/bodies/body_queries.hpp
@@ -18,7 +18,7 @@
 
 #include <silkworm/db/datastore/snapshots/basic_queries.hpp>
 
-#include "body_snapshot.hpp"
+#include "body_segment.hpp"
 
 namespace silkworm::snapshots {
