+++ silkworm/db/blocks/bodies/body_segment.cpp
@@ -14,16 +14,19 @@
    limitations under the License.
 */
 
-#pragma once
+#include "body_segment.hpp"
 
-#include "rec_split_index/index.hpp"
-#include "segment/snapshot_reader.hpp"
+#include <silkworm/infra/common/decoding_exception.hpp>
 
 namespace silkworm::snapshots {
 
-struct SegmentAndIndex {
-    const SegmentFileReader& segment;
-    const Index& index;
-};
+void encode_word_from_body(Bytes& word, const BlockBodyForStorage& body) {
+    word = body.encode();
+}
+
+void decode_word_into_body(ByteView word, BlockBodyForStorage& body) {
+    const auto result = decode_stored_block_body(word, body);
+    success_or_throw(result, "decode_word_into_body: decode_stored_block_body error");
+}
 
 }  // namespace silkworm::snapshots
