+++ silkworm/db/datastore/snapshots/segment/segment_writer.hpp
