+++ silkworm/db/freezer.cpp
@@ -30,7 +30,7 @@
 #include "blocks/headers/header_snapshot_freezer.hpp"
 #include "datastore/snapshot_freezer.hpp"
 #include "datastore/snapshots/common/snapshot_path.hpp"
-#include "datastore/snapshots/segment/snapshot_writer.hpp"
+#include "datastore/snapshots/segment/segment_writer.hpp"
 #include "datastore/snapshots/snapshot_bundle.hpp"
 #include "prune_mode.hpp"
 #include "transactions/txn_snapshot_freezer.hpp"
