+++ silkworm/db/snapshot_recompress.cpp
@@ -19,10 +19,10 @@
 
 #include <silkworm/infra/common/directories.hpp>
 
-#include "blocks/bodies/body_snapshot.hpp"
-#include "blocks/headers/header_snapshot.hpp"
+#include "blocks/bodies/body_segment.hpp"
+#include "blocks/headers/header_segment.hpp"
 #include "datastore/snapshots/common/snapshot_path.hpp"
-#include "transactions/txn_snapshot.hpp"
+#include "transactions/txn_segment.hpp"
 
 namespace silkworm::snapshots {
