+++ silkworm/db/transactions/txn_index.cpp
@@ -17,9 +17,9 @@
 #include "txn_index.hpp"
 
 #include <silkworm/db/blocks/bodies/body_txs_amount_query.hpp>
-#include <silkworm/db/datastore/snapshots/segment/snapshot_reader.hpp>
+#include <silkworm/db/datastore/snapshots/segment/segment_reader.hpp>
 
-#include "txn_snapshot_word_serializer.hpp"
+#include "txn_segment_word_serializer.hpp"
 
 namespace silkworm::snapshots {
