+++ silkworm/db/transactions/txn_segment_word_serializer.hpp
